Move CaptureResAeff
