+++ marxs/design/tolerancing.py
@@ -11,15 +11,11 @@ from astropy import table
 import astropy.units as u
 from marxs.simulator import Sequence
 from marxs.design.uncertainties import generate_facet_uncertainty as genfacun
-from marxs.analysis.gratings import (resolvingpower_from_photonlist,
-                                 effectivearea_from_photonlist)
-from marxs.analysis.gratings import AnalysisError
 
 __all__ = ['oneormoreelements',
            'wiggle', 'moveglobal', 'moveindividual', 'moveelem',
            'varyperiod', 'varyorderselector', 'varyattribute',
            'run_tolerances', 'run_tolerances_for_energies',
-           'CaptureResAeff',
            'generate_6d_wigglelist', 'reset_6d',
            'select_1dof_changed',
            'plot_wiggle', 'load_and_plot',
@@ -283,184 +279,6 @@ def run_tolerances(photons_in, instrum, wigglefunc, wiggleparts,
         out.append(cpars)
 
     return out
-
-
-class CaptureResAeff():
-    '''Capture resolving power and effective area for a tolerancing simulation.
-
-    Instances of this class can be called with a photon list for a tolerancing
-    simulation. The photon list will be analysed for resolving power and
-    effective area in a number of relevant orders.
-
-    This is implemented as a class and not a simple function. When the class is
-    initialized a number of parameters that are true for any to the
-    analysis (e.g. the names of certain columns) are set and saved in the class
-    instance.
-
-    The implementation of this class is geared towards instruments with
-    gratings but can also serve as an example how a complex analysis that
-    derives several different parameters can be implemented.
-
-    Results for effective area and resolving power are reported on a per order
-    basis and also summarized for all grating orders combined and the zeroth
-    order separately.
-
-    Inherit form this class to customize the filters applied to the photon list
-    before effective area and resolving power are calculated.
-
-    Parameters
-    ----------
-    A_geom : number
-        Geometric area of aperture for the simulations that this instance
-        will analyze.
-    order_col : string
-        Column names for grating orders
-    orders : array
-        Order numbers to consider in the analysis
-    dispersion_coord : string
-        Dispersion coordinate for
-        `marxs.analysis.gratings.resolvingpower_from_photonlist`. Any photons
-        that have non-finite values or are masked in this column will be ignored
-        for the purpose of the resolving power calculation.
-    '''
-    def __init__(self, A_geom=1, order_col='order',
-                 orders=np.arange(-10, 11),
-                 dispersion_coord='det_x'):
-        self.A_geom = A_geom
-        self.order_col = order_col
-        self.orders = np.asanyarray(orders)
-        self.dispersion_coord = dispersion_coord
-
-    def aeff_filter(self, photons):
-        '''Filter photon list before calculating the effective area.
-
-        Here: No-op (all photons are used.)
-
-        Parameters
-        ----------
-        photons : `astropy.table.Table`
-            Photon list.
-
-        Returns
-        -------
-        ind : numpy array
-           array of boolean values that can be used to index ``photons`` and select
-           a sub-set to be used to calcualte the effective area.
-        '''
-        return np.ones(len(photons), dtype=bool)
-
-    def res_filter(self, photons):
-        ind = (np.isfinite(photons[self.dispersion_coord]) &
-                (photons['probability'] > 0))
-        if hasattr(photons[self.dispersion_coord], 'mask'):
-            ind = ind & ~photons[self.dispersion_coord].mask
-        return ind
-
-    def __call__(self, photons):
-        '''Calculate Aeff and R for an input photon list.
-
-        Parameters
-        ----------
-        photons : `astropy.table.Table`
-            Photon list.
-
-        Returns
-        -------
-        result : dict
-            Dictionary with per-order Aeff and R, as well as values
-            summed over all grating orders.
-        '''
-        ind = self.aeff_filter(photons)
-        aeff = effectivearea_from_photonlist(photons[ind], self.orders, len(photons),
-                                             self.A_geom, self.order_col)
-        try:
-            ind = self.res_filter(photons)
-            res, pos, std = resolvingpower_from_photonlist(photons[ind], self.orders,
-                                                           col=self.dispersion_coord,
-                                                           zeropos=None,
-                                                           ordercol=self.order_col)
-        except AnalysisError:
-            # Something did not work, e.g. too few photons to find zeroth order
-            res = np.nan * np.ones(len(self.orders))
-
-        disporders = self.orders != 0
-        # The following lines work for an empty photon list, too.
-        aeffgrat = np.sum(aeff[disporders])
-        aeff0 = np.sum(aeff[~disporders])
-
-        # Division by 0 causes more nans, so filter those out
-        # Also, res is nan if less than 20 photons are detected
-        # so we need to filter those out, too.
-        ind = disporders & (aeff > 0) & np.isfinite(res)
-        if ind.sum() == 0:  # Dispersed spectrum misses detector
-            avggratres = np.nan
-        else:
-            avggratres = np.average(res[ind],
-                                    weights=aeff[ind] / aeff[ind].sum())
-        return {'Aeff0': aeff0, 'Aeffgrat': aeffgrat, 'Aeff': aeff,
-                'Rgrat': avggratres, 'R': res}
-
-
-class CaptureResAeff_CCDgaps(CaptureResAeff):
-    '''Capture resolving power and effective area for a tolerancing simulation.
-
-    Unlike `CaptureResAeff` this objects is set up to take into account CCD gaps.
-    For the resolving power calculation, one wants to ignore CCD gaps, because they
-    might make the observed LSF artificially small - if only half of the LSF falls
-    on a CCD, and the other photons are not counted, the LSF will appear only
-    half as wide. In contrast, for the effective area calcalution, we *do* want to
-    ignore photons that miss a CCD.
-
-    Here, we following approach is taken: This class ovverrides
-    `~marxs.design.tolerancing.CaptureResAeff.aeff_filter`.
-    All photons with a probability >0 are presumed ot reach the focal plane.
-    For the effective area, an additional
-    filter is applied: Only photons with `photons['aeff_filter_col'] > 0` will
-    be used. Typically, this could be the CCD_ID.
-
-    Parameters
-    ----------
-    A_geom : number
-        Geometric area of aperture for the simulations that this instance
-        will analyze.
-    order_col : string
-        Column names for grating orders
-    orders : array
-        Order numbers to consider in the analysis
-    dispersion_coord : string
-        Dispersion coordinate for
-        `marxs.analysis.gratings.resolvingpower_from_photonlist`. Any photons
-        that have non-finite values or are masked in this column will be ignored
-        for the purpose of the resolving power calculation.
-    aeff_filter_col : string
-        Only photons where the value in this column is larger than 0 will be used
-        for the calculation of the effective area.
-    '''
-    def __init__(self, A_geom=1, order_col='order',
-                 orders=np.arange(-10, 11),
-                 dispersion_coord='det_x', aeff_filter_col='CCD_ID'):
-        super().__init__(A_geom=A_geom, order_col=order_col,
-                 orders=orders,
-                 dispersion_coord=dispersion_coord)
-        self.aeff_filter_col = aeff_filter_col
-
-    def aeff_filter(self, photons):
-        '''Filter photon list before calculating the effective area.
-
-        Only photons with ``photons[aeff_filter_col] >= 0`` are used.
-
-        Parameters
-        ----------
-        photons : `astropy.table.Table`
-            Photon list.
-
-        Returns
-        -------
-        ind : numpy array
-           array of boolean values that can be used to index ``photons`` and select
-           a sub-set to be used to calcualte the effective area.
-        '''
-        return photons[self.aeff_filter_col] >= 0
 
 
 def run_tolerances_for_energies(source, energies,
